requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
certifi
# Task Queue
celery==5.3.4
redis==5.0.1

# AI/ML - LLM Providers (install based on your chosen provider)
anthropic==0.17.0

openai==1.12.0
google-generativeai==0.3.2

# LangChain & LangGraph
#langchain==0.1.0
#langchain-anthropic==0.1.1
#langgraph==0.0.40

langchain
langchain-anthropic
langchain-community
langgraph

# Machine Learning
scikit-learn==1.2.2

# scikit-learn==1.4.0
numpy==1.26.3

# Databases
pymongo==4.6.1
motor==3.3.2
neo4j==5.15.0
redis==5.0.1

# GitHub Integration
PyGithub==2.1.1
GitPython==3.1.41

# HTTP Clients
httpx==0.26.0
requests==2.31.0

# Code Analysis
tree-sitter==0.20.4
libcst==1.1.0

# Observability
arize-phoenix==2.1.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic-settings==2.1.0

# Security & Encryption
cryptography==42.0.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
ruff==0.1.13
mypy==1.8.0
